Rename project back to pyipatcher
