+++ pyproject.toml
@@ -1,18 +1,17 @@
 [tool.poetry]
-name = "patchfinder64"
-version = "1.0.0"
-description = "iOS ARM64 kernel patchfinder"
+name = "pyipatcher"
+version = "1.0.1"
+description = "iOS ARM64 patcher"
 authors = ["mini_exploit <61931266+Mini-Exploit@users.noreply.github.com>"]
 license = "GPL-3.0-only"
 readme = "README.md"
 repository = "https://github.com/Mini-Exploit/pyipatcher"
 
 [tool.poetry.scripts]
-kernelpatcher = "patchfinder64.kernelpatcher:main"
+kernelpatcher = "pyipatcher.kernelpatcher:main"
 
 [tool.poetry.dependencies]
 python = "^3.7"
-m1n1Exception = "^2.7"
 click = "^8.1.3"
 
 [tool.poetry.group.dev.dependencies]
@@ -26,7 +25,7 @@ skip-string-normalization = true
 
 [tool.isort]
 profile = "black"
-src_paths = ["patchfinder64"]
+src_paths = ["pyipatcher"]
 
 [tool.pytest.ini_options]
 # TODO: Add proper tests
